moved import and module in seperate module_declaration syntax
