+++ lang/fun/src/parser/fun.lalrpop
@@ -2,6 +2,7 @@ use crate::syntax::*;
 use crate::syntax::arguments::*;
 use crate::syntax::context::*;
 use crate::syntax::declarations::*;
+use crate::syntax::module_declarations::*;
 use crate::syntax::names::*;
 use crate::syntax::program::*;
 use crate::syntax::terms::*;
@@ -483,14 +484,17 @@ Codata: Codata = {
 }
 
 Declaration: Declaration = {
-    <module: Module> => module.into(),
-    <import: Import> => import.into(),
     <pdef: PDef> => pdef.into(),
     <def: Def> => def.into(),
     <data: Data> => data.into(),
     <codata: Codata> => codata.into()
 }
 
+ModuleDeclaration: ModuleDeclaration = {
+    <module: Module> => module.into(),
+    <import: Import> => import.into(),
+}
+
 pub Prog: Program = {
-    <declarations: Declaration*> => Program { declarations }
+    <module_declarations: ModuleDeclaration*> <declarations: Declaration*> => Program { module_declarations, declarations }
 }
