+++ lang/fun/src/syntax/declarations/mod.rs
@@ -6,15 +6,11 @@ use printer::*;
 pub mod codata;
 pub mod data;
 pub mod def;
-pub mod module;
-pub mod import;
 pub mod pdef;
 
 pub use codata::*;
 pub use data::*;
 pub use def::*;
-pub use module::*;
-pub use import::*;
 pub use pdef::*;
 
 /// This enum encodes whether a user-declared type is a data or codata type.
@@ -30,8 +26,6 @@ pub enum Polarity {
 /// or top-level function [`Def`]initions.
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub enum Declaration {
-    Module(Module),
-    Import(Import),
     Data(Data),
     Codata(Codata),
     PDef(PDef),
@@ -41,8 +35,6 @@ pub enum Declaration {
 impl Print for Declaration {
     fn print<'a>(&'a self, cfg: &PrintCfg, alloc: &'a Alloc<'a>) -> Builder<'a> {
         match self {
-            Declaration::Module(module) => module.print(cfg, alloc),
-            Declaration::Import(import) => import.print(cfg, alloc),
             Declaration::Def(def) => def.print(cfg, alloc),
             Declaration::PDef(pdef) => pdef.print(cfg, alloc),
             Declaration::Data(data) => data.print(cfg, alloc),
